requirements.txt
# YouTube Agent Roma - Dependencies

# Core dependencies
requests>=2.31.0
pyyaml>=6.0
click>=8.1.0

# YouTube API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web scraping (if needed)
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Async support
aiohttp>=3.8.0
asyncio

# Configuration and logging
python-dotenv>=1.0.0
loguru>=0.7.0

# Web API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

